# packages/api-types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mini Hedge Fund Desk",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/agent-token": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Create Agent Token",
        "description": "Issue a JWT for an LLM agent.\n\nRequires an authenticated user with ``funds:manage`` permission.\nThe agent token is scoped to the caller's fund and carries the\ndelegating user's ID for audit. The agent cannot be granted more\npermissions than the delegating user holds.",
        "operationId": "create_agent_token_api_v1_auth_agent_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audit/verify": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Verify Audit Integrity",
        "description": "Run integrity verification on the audit log hash chain.",
        "operationId": "verify_audit_integrity_api_v1_audit_verify_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/tokens/revoke": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Revoke Token",
        "description": "Revoke a specific token by its JTI. Requires platform admin permission.",
        "operationId": "revoke_token_api_v1_platform_tokens_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/users/{user_id}/revoke-tokens": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Revoke User Tokens",
        "description": "Revoke all tokens for a user. Requires platform admin permission.",
        "operationId": "revoke_user_tokens_api_v1_platform_users__user_id__revoke_tokens_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeUserTokensResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentTokenRequest": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "default": [
              "viewer"
            ]
          }
        },
        "type": "object",
        "required": [
          "agent_name"
        ],
        "title": "AgentTokenRequest"
      },
      "AgentTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type"
          },
          "fund_slug": {
            "type": "string",
            "title": "Fund Slug"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "actor_type",
          "fund_slug",
          "roles"
        ],
        "title": "AgentTokenResponse"
      },
...
      "AuditVerifyResponse": {
        "properties": {
          "is_valid": {
            "type": "boolean",
            "title": "Is Valid"
          },
          "records_checked": {
            "type": "integer",
            "title": "Records Checked"
          },
          "first_broken_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Broken Link"
          }
        },
        "type": "object",
        "required": [
          "is_valid",
          "records_checked"
        ],
        "title": "AuditVerifyResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RevokeTokenRequest": {
        "properties": {
          "jti": {
            "type": "string",
            "title": "Jti"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "jti",
          "expires_at"
        ],
        "title": "RevokeTokenRequest"
      },
      "RevokeTokenResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "revoked"
          },
          "jti": {
            "type": "string",
            "title": "Jti"
          }
        },
        "type": "object",
        "required": [
          "jti"
        ],
        "title": "RevokeTokenResponse"
      },
      "RevokeUserTokensResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "revoked"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "RevokeUserTokensResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}